pyproject.toml
[project]
name = "spotify-analytics"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.2",
    "dj-rest-auth[with-social]>=7.0.1",
    "django>=5.2.9",
    "django-allauth[socialaccount]>=65.13.1",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "geoip2>=5.2.0",
    "psycopg2-binary>=2.9.11",
    "tenacity>=9.1.2",
]
